Fixed crucial bug in nearest neighbors search.
The nearest neighbors search works by querying the kd-tree for a number
of neighbors (starting from 2) until a neighbor with a non-zero
distance and satisfying the Theiler window criterion is found.
Sometimes, a point will have two or more neighbors with the same
distance.  In such cases, the kd-tree need not return the same list of
neighbors when it's asked for more neighbors.  This can lead to more
than the required number of points satisfying the conditions.  We have
to check this and take only the required number of points.