Normalize the RMSD in IAAFT surrogates.
Now I'm calculating the RMSD between the absolute squares of the Fourier
amplitudes instead of calculating it between just the absolute values of
the Fourier amplitudes.  I'm also normalizing it by the mean of the
power spectrum.  Of course, none of this actually helps, but it gives me
an ounce of satisfaction to use "standard" statistics.